refactor: #347
TicketingOrderClient를 인터페이스와 API 구현체로 분리
